requirements.txt
alembic==1.9.1
autopep8==2.0.0
black==22.10.0
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-CKEditor==0.4.6
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==1.1.3.post0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pycodestyle==2.9.1
PyJWT==2.6.0
python-dotenv==0.21.0
python-slugify==7.0.0
SQLAlchemy==1.4.42
text-unidecode==1.3
tomli==2.0.1
Werkzeug==2.2.2
WTForms==3.0.1
